requirements.txt
black==22.12.0
anyio==3.6.1
cffi==1.15.1
click==8.1.3
cryptography==43.0.0
fastapi==0.95.2
flake8==6.0.0
h11==0.13.0
idna==3.4
isort==5.11.4
pre-commit==2.21.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.9.0
python-dotenv==0.21.0
secure==0.3.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.3.0
uvicorn==0.18.3
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
